+++ examples/maven-basic/src/test/java/spec/com/example/CalculatorSpec.java
@@ -2,6 +2,6 @@ package spec.com.example;
 
 public class CalculatorSpec extends CalculatorSpecSupport {
     public void it_adds_two_numbers() {
-        match(subject().add(2, 3)).shouldReturn(5);
+        add(2, 3).shouldReturn(5);
     }
 }